requirements.txt
pandas
numpy
pyarrow
requests
boto3
duckdb==1.2.2
python-dotenv
asset-model-data-storage
flask